requirements.txt
streamlit
google-generativeai
python-dotenv
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-google-genai
langchain-text-splitters
faiss-cpu
pypdf